requirements.txt
fastapi
pydantic
uvicorn
requests
pymongo
python-dotenv
numpy>=1.22.4,<2.0.0
opencv-python
reedsolo
scipy
python-multipart
cryptography 
passlib[bcrypt]==1.7.4
